Change rake_task_test assertions to not be dependent on order
Probably due to the recent inclusion of a seed parameter for minitest, the assertions in this test sometimes failed due to the test output being in a different order to the expectations. Now the order doesn't matter so it should always pass as long as all the expected results did happen.